src/main/java/com/bridgelabz/statecensusanalyserproblem/StateCensusAnalyser.java
package com.bridgelabz.statecensusanalyserproblem;

import com.opencsv.bean.CsvToBean;
import com.opencsv.bean.CsvToBeanBuilder;

import java.io.IOException;
import java.io.Reader;
import java.nio.file.Files;
import java.nio.file.NoSuchFileException;
import java.nio.file.Paths;
import java.util.Iterator;

public class StateCensusAnalyser {


    private String STATE_CODE_CSV_FILE_PATH = "/home/admin265/IdeaProjects/StateCensusAnalyserProblem/src/main/resources/StateCode.csv";

    public StateCensusAnalyser(String STATE_CODE_CSV_FILE_PATH) {
        this.STATE_CODE_CSV_FILE_PATH=STATE_CODE_CSV_FILE_PATH;
    }

    public int readStateData() throws CsvException {
        int count = 0;
        try (Reader reader = Files.newBufferedReader(Paths.get(STATE_CODE_CSV_FILE_PATH));) {
            CsvToBean<CSVStates> csvToBean = new CsvToBeanBuilder(reader)
                    .withIgnoreLeadingWhiteSpace(true)
                    .withType(CSVStates.class)
                    .build();
            Iterator<CSVStates> stateIterator = csvToBean.iterator();
            while (stateIterator.hasNext()) {
                CSVStates state = stateIterator.next();
                count++;
            }
        } catch (NoSuchFileException e) {
            throw new CsvException(CsvException.ExceptionType.FILE_NOT_FOUND, "Such type file doesn't exist", e.getCause());
        } catch (IOException e) {
            e.printStackTrace();
        }
        return count;
    }
}